Fix tests from moving serializable namespaces and name of the ndoe type
