+++ test/unit/operations/extract/basic/falsyValues.ts
@@ -1,6 +1,5 @@
 import { extract } from '../../../../../src/operations/extract';
-import { Serializeable } from '../../../../../src/primitive';
-import { StaticNodeId } from '../../../../../src/schema';
+import { Serializeable, StaticNodeId } from '../../../../../src/schema';
 import { createSnapshot } from '../../../../helpers';
 
 const { QueryRoot: QueryRootId } = StaticNodeId;
@@ -21,7 +20,7 @@ describe.skip(`operations.extract`, () => {
     it(`extracts JSON serializable object`, () => {
       expect(extractResult).to.deep.eq({
         [QueryRootId]: {
-          nodeSnapshotType: Serializeable.NodeSnapshotType.EntitySnapshot,
+          type: Serializeable.NodeSnapshotType.EntitySnapshot,
           data: { null: null, false: false, zero: 0, string: '' },
         },
       });
